RFC 023 phase 2 & 3
